Move invocation api, implement JVM_FindClassFromBootLoader
